# frontend/apps/site/openapi-spec.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "CloudCrafter API",
		"version": "v1"
	},
	"paths": {
...
		"/api/Teams": {
			"get": {
				"tags": ["Teams"],
				"operationId": "GetMyTeams",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/SimpleTeamDto"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/SimpleTeamDto"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/SimpleTeamDto"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Teams"],
				"operationId": "CreateTeam",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTeamCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTeamCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTeamCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "string",
									"format": "uuid"
								}
							},
							"application/json": {
								"schema": {
									"type": "string",
									"format": "uuid"
								}
							},
							"text/json": {
								"schema": {
									"type": "string",
									"format": "uuid"
								}
							}
						}
					}
				}
			}
		},
		"/api/Teams/{teamId}": {
			"get": {
				"tags": ["Teams"],
				"operationId": "GetTeamMembers",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "Filters",
						"in": "query",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/components/schemas/FilterCriterea"
							}
						}
					},
					{
						"name": "Page",
						"in": "query",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "PageSize",
						"in": "query",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedListOfTeamMemberDto"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedListOfTeamMemberDto"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedListOfTeamMemberDto"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Teams"],
				"operationId": "RenameTeam",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RenameTeamCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/RenameTeamCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/RenameTeamCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			},
			"delete": {
				"tags": ["Teams"],
				"operationId": "DeleteTeam",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/api/Teams/{teamId}/invite": {
			"post": {
				"tags": ["Teams"],
				"operationId": "InviteUserToTeam",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InviteUserBody"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/InviteUserBody"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/InviteUserBody"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/api/Teams/{teamId}/remove": {
			"delete": {
				"tags": ["Teams"],
				"operationId": "RemoveUserFromTeam",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RemoveMemberBody"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/RemoveMemberBody"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/RemoveMemberBody"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CreateTeamCommand": {
				"required": ["name"],
				"type": "object",
				"properties": {
					"name": {
						"minLength": 1,
						"type": "string"
					}
				}
			},
...
			"FilterCriterea": {
				"required": ["propertyName", "operator"],
				"type": "object",
				"properties": {
					"propertyName": {
						"type": "string"
					},
					"operator": {
						"$ref": "#/components/schemas/FilterOperatorOption"
					},
					"value": {
						"type": "string",
						"nullable": true
					}
				}
			},
			"FilterOperatorOption": {
				"enum": [
					"Equal",
					"NotEqual",
					"GreaterThan",
					"GreaterThanOrEqual",
					"LessThan",
					"LessThanOrEqual",
					"Contains"
				]
			},
...
			"InviteUserBody": {
				"required": ["email"],
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					}
				}
			},
...
			"PaginatedListOfTeamMemberDto": {
				"required": ["page", "totalPages", "totalItems", "result"],
				"type": "object",
				"properties": {
					"page": {
						"type": "integer",
						"format": "int32"
					},
					"totalPages": {
						"type": "integer",
						"format": "int32"
					},
					"totalItems": {
						"type": "integer",
						"format": "int32"
					},
					"result": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/TeamMemberDto"
						}
					}
				}
			},
...
			"RemoveMemberBody": {
				"required": ["email"],
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					}
				}
			},
			"RenameTeamCommand": {
				"required": ["name"],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					}
				}
			},
...
			"SimpleTeamDto": {
				"required": ["id", "name"],
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					}
				}
			},
...
			"TeamMemberDto": {
				"required": ["id", "email", "fullName"],
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"email": {
						"type": "string",
						"nullable": true
					},
					"fullName": {
						"type": "string",
						"nullable": true
					}
				}
			},
...
		}
	},
...
}